requirements.txt
requests==2.31.0
python-dotenv==1.0.0
transformers==4.36.2
huggingface-hub==0.20.3
torch==2.2.0
sentencepiece
